Update purescript-redis to the correct commit id
The previously used commit id was stale. The update requires that we
move all CacheConn references to SimpleConn. This does raise the
question of how we might want to do Cluster support as that introduces
the need to express the connection type as a typeclass (CacheConn a),
but that might require dramatic changes, or dropping Redis support
entirely from Presto Backend.